+++ tests/DoctrineAuditBundle/BaseTest.php
@@ -4,8 +4,8 @@ namespace DH\DoctrineAuditBundle\Tests;
 
 use DH\DoctrineAuditBundle\AuditConfiguration;
 use DH\DoctrineAuditBundle\EventSubscriber\AuditSubscriber;
-use DH\DoctrineAuditBundle\EventSubscriber\CreateSchemaListener;
 use DH\DoctrineAuditBundle\Helper\AuditHelper;
+use DH\DoctrineAuditBundle\Helper\UpdateHelper;
 use DH\DoctrineAuditBundle\Manager\AuditManager;
 use DH\DoctrineAuditBundle\Reader\AuditReader;
 use DH\DoctrineAuditBundle\User\TokenStorageUserProvider;
@@ -59,6 +59,7 @@ abstract class BaseTest extends TestCase
         $this->getEntityManager();
         $this->getSchemaTool();
         $this->setUpEntitySchema();
+        $this->setUpAuditSchema();
     }
 
     /**
@@ -67,6 +68,7 @@ abstract class BaseTest extends TestCase
      */
     protected function tearDown(): void
     {
+        $this->tearDownAuditSchema();
         $this->tearDownEntitySchema();
         $this->em = null;
         $this->schemaTool = null;
@@ -121,6 +123,45 @@ abstract class BaseTest extends TestCase
                 $statement = $em->getConnection()->prepare($query);
                 $statement->execute();
             } catch (\Exception $e) {
+            }
+        }
+    }
+
+    protected function setUpAuditSchema(): void
+    {
+        $configuration = $this->getAuditConfiguration();
+        $helper = new AuditHelper($configuration);
+        $manager = new AuditManager($configuration, $helper);
+        $reader = $this->getReader($this->getAuditConfiguration());
+
+        $updater = new UpdateHelper($manager, $reader);
+        $updater->updateAuditSchema();
+    }
+
+    protected function tearDownAuditSchema(): void
+    {
+        $configuration = $this->getAuditConfiguration();
+        $em = $configuration->getEntityManager();
+        $schemaManager = $em->getConnection()->getSchemaManager();
+        $schema = $schemaManager->createSchema();
+        $fromSchema = clone $schema;
+
+        $tables = $schemaManager->listTables();
+        foreach ($tables as $table) {
+            $regex = '#^'.$configuration->getTablePrefix().'.*'.$configuration->getTableSuffix().'$#';
+            if (preg_match($regex, $table->getName())) {
+                $schema->dropTable($table->getName());
+            }
+        }
+
+        $sqls = $fromSchema->getMigrateToSql($schema, $schemaManager->getDatabasePlatform());
+
+        foreach ($sqls as $sql) {
+            try {
+                $statement = $em->getConnection()->prepare($sql);
+                $statement->execute();
+            } catch (\Exception $e) {
+                // something bad happened here :/
             }
         }
     }
@@ -197,7 +238,6 @@ abstract class BaseTest extends TestCase
             }
         }
         $evm->addEventSubscriber(new AuditSubscriber($this->auditManager));
-        $evm->addEventSubscriber(new CreateSchemaListener($this->auditManager));
         $evm->addEventSubscriber(new Gedmo\SoftDeleteable\SoftDeleteableListener());
 
         return $this->em;