+++ tests/DoctrineAuditBundle/IssueTest.php
@@ -16,7 +16,6 @@ use DH\DoctrineAuditBundle\Tests\Fixtures\Issue40\DieselCase;
  * @covers \DH\DoctrineAuditBundle\DBAL\AuditLogger
  * @covers \DH\DoctrineAuditBundle\DBAL\AuditLoggerChain
  * @covers \DH\DoctrineAuditBundle\EventSubscriber\AuditSubscriber
- * @covers \DH\DoctrineAuditBundle\EventSubscriber\CreateSchemaListener
  * @covers \DH\DoctrineAuditBundle\Helper\AuditHelper
  * @covers \DH\DoctrineAuditBundle\Helper\DoctrineHelper
  * @covers \DH\DoctrineAuditBundle\Helper\UpdateHelper
@@ -175,5 +174,6 @@ final class IssueTest extends BaseTest
         ]);
 
         $this->setUpEntitySchema();
+        $this->setUpAuditSchema();
     }
 }